ECO-871. Refactoring of express checkout is finished.
 * ExpressCheckoutHandler has been removed.
 * Logic from ExpressCheckoutHandler has been moved to steps.
