+++ src/SprykerEco/Yves/Payone/Checkout/Process/Steps/LoadExpressCheckoutDetailsStep.php
@@ -1,32 +1,49 @@
 <?php
+
 /**
- * Created by PhpStorm.
- * User: sikachev
- * Date: 10/13/17
- * Time: 5:37 PM
+ * MIT License
+ * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
  */
 
 namespace SprykerEco\Yves\Payone\Checkout\Process\Steps;
 
-
+use Spryker\Client\Cart\CartClientInterface;
 use Spryker\Shared\Kernel\Transfer\AbstractTransfer;
 use Spryker\Yves\StepEngine\Dependency\Step\AbstractBaseStep;
-use SprykerEco\Yves\Payone\Handler\ExpressCheckoutHandlerInterface;
+use SprykerEco\Client\Payone\PayoneClientInterface;
+use SprykerEco\Yves\Payone\Handler\ExpressCheckout\QuoteHydratorInterface;
 use Symfony\Component\HttpFoundation\Request;
 
 class LoadExpressCheckoutDetailsStep extends AbstractBaseStep
 {
 
-    protected $expressCheckoutHandler;
+    /**
+     * @var \Spryker\Client\Cart\CartClientInterface
+     */
+    protected $cartClient;
+
+    /**
+     * @var \SprykerEco\Client\Payone\PayoneClientInterface
+     */
+    protected $payoneClient;
+
+    /**
+     * @var \SprykerEco\Yves\Payone\Handler\ExpressCheckout\QuoteHydratorInterface
+     */
+    protected $quoteHydrator;
 
     public function __construct(
-        ExpressCheckoutHandlerInterface $expressCheckoutHandler,
+        CartClientInterface $cartClient,
+        PayoneClientInterface $payoneClient,
+        QuoteHydratorInterface $quoteHydrator,
         $stepRoute,
         $escapeRoute
     )
     {
         parent::__construct($stepRoute, $escapeRoute);
-        $this->expressCheckoutHandler = $expressCheckoutHandler;
+        $this->cartClient = $cartClient;
+        $this->payoneClient = $payoneClient;
+        $this->quoteHydrator = $quoteHydrator;
     }
 
     /**
@@ -63,9 +80,12 @@ class LoadExpressCheckoutDetailsStep extends AbstractBaseStep
      */
     public function execute(Request $request, AbstractTransfer $dataTransfer)
     {
-        //TODO: check why the second request fails or cache the first request.
-        $this->expressCheckoutHandler
-            ->loadPaypalExpressCheckoutDetails();
+        $quoteTransfer = $this->cartClient->getQuote();
+        $details = $this->payoneClient->getPaypalExpressCheckoutDetails($quoteTransfer);
+        $quoteTransfer = $this->quoteHydrator->getHydratedQuote($quoteTransfer, $details);
+        $this->cartClient->storeQuote($quoteTransfer);
+
+        return $quoteTransfer;
 
     }
