+++ src/SprykerEco/Yves/Payone/Theme/default/express-checkout/success.twig
@@ -12,3 +12,6 @@
         </div>
     </div>
 {% endblock %}
+
+{% block footer %}
+{% endblock %}
